# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Device Farm API",
    "version": "3.0.0",
    "description": "Self-hosted test execution platform for Apple Silicon Macs"
  },
  "components": {
    "schemas": {
...
      "Hook": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "event": {
            "type": "string",
            "enum": [
              "device.booted",
              "device.shutdown",
              "test.before",
              "test.after"
            ]
          },
          "command": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096
          },
          "platform": {
            "default": "all",
            "type": "string",
            "enum": [
              "android",
              "ios",
              "all"
            ]
          },
          "timeoutMs": {
            "default": 30000,
            "type": "integer",
            "minimum": 1000,
            "maximum": 300000
          },
          "failOnError": {
            "default": false,
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "event",
          "command",
          "platform",
          "timeoutMs",
          "failOnError",
          "enabled"
        ],
        "additionalProperties": false,
        "description": "Registered lifecycle hook"
      },
      "HookConflict": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "number",
            "const": 409
          },
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "additionalProperties": false,
        "description": "RFC 7807 conflict body for duplicate hook names"
      },
...
    }
  },
  "paths": {
...
    "/api/hooks": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "event": {
                    "type": "string",
                    "enum": [
                      "device.booted",
                      "device.shutdown",
                      "test.before",
                      "test.after"
                    ]
                  },
                  "command": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4096
                  },
                  "platform": {
                    "default": "all",
                    "type": "string",
                    "enum": [
                      "android",
                      "ios",
                      "all"
                    ]
                  },
                  "timeoutMs": {
                    "default": 30000,
                    "type": "integer",
                    "minimum": 1000,
                    "maximum": 300000
                  },
                  "failOnError": {
                    "default": false,
                    "type": "boolean"
                  },
                  "enabled": {
                    "default": true,
                    "type": "boolean"
                  }
                },
                "required": [
                  "name",
                  "event",
                  "command"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hook"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookConflict"
                }
              }
            }
          }
        }
      }
    },
    "/api/hooks/{name}": {
      "put": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/hooks/{name}/test": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}